% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_transcript_info.R
\name{load_biomart}
\alias{load_biomart}
\title{Load Specific Ensembl Biomart}
\usage{
load_biomart(GRCh = c("38", "37"))
}
\arguments{
\item{GRCh}{reference assembly ('38' or '37')}
}
\value{
By default, returns a gene database from the latest ensembl hg38/hg37 biomart
}
\description{
Load Specific Ensembl Biomart
}
