% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_collapse_duplicate}
\alias{probes_collapse_duplicate}
\title{Collapse duplicates}
\usage{
probes_collapse_duplicate(df_mut)
}
\arguments{
\item{df_mut}{output from \code{\link[=probes_construct]{probes_construct()}}}
}
\value{
Duplicate probes collapsed into single probes (data.frame)
}
\description{
Collapse duplicates
}
