% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_collapse_duplicates}
\alias{probes_collapse_duplicates}
\title{Collapse duplicates}
\usage{
probes_collapse_duplicates(df_mut)
}
\arguments{
\item{df_mut}{output of probes_construct}
}
\value{
data.frame with duplicate probes collapsed
}
\description{
Collapse duplicates
}
