% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_construct}
\alias{probes_construct}
\title{Construct Probes}
\usage{
probes_construct(
  df_mut,
  ensembl,
  probe_size = 51,
  probe_type = c("mRNA_no_U", "mRNA")
)
}
\arguments{
\item{df_mut}{data frame of mutations (produced from probes_read_vep_txt) (data.frame)}

\item{ensembl}{biomart from which to fetch transcript (cDNA)  sequences. Define using \code{\link[=load_biomart]{load_biomart()}}}

\item{probe_size}{How large should WT and Mutant probes be (number)}

\item{probe_type}{which type of probe sequence to return (mRNA_no_U, mRNA)}
}
\value{
data.frame describing 1 probe per mutation-transcript isoform combination.
Try piping into probes_collapse_duplicates to cleanup
}
\description{
Construct Probes
}
\details{
There are many ways to build probes. Our approach is as follows.
\enumerate{
\item Define a fixed probe_size
\item Dynamically calculate the number of bases upstream and downstream based on total probe size.
Done separately for WT and Mutant since mutant seq may have insertions
}

Note where mutation and probe size does not allow for identical sizes of upstream and downstream flanking region,
we always place the extra base on the upstream size

Key formulas used:

Ts = Total probe size = defined by probe_size

L = mutation length (SNV = 1)

Uwt = Upstream bases (wild type) = floor((Ts - 1)/2)

Dwt = Downstream bases (wild type) = ceiling((Ts - 1)/2)

Umut = Upstream bases (mutant probe) = floor((Ts-L)/2)

Dmut = Downstream bases (mutant probe) = ceiling((Ts-L)/2)
}
