% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_construct_fusion_sequence}
\alias{probes_construct_fusion_sequence}
\title{Fusion Sequence to Probe}
\usage{
probes_construct_fusion_sequence(fusion, probe_size = 51)
}
\arguments{
\item{fusion}{a vector of strings where each string is a fusion sequence with breakpoint indicated using |. Should be in mRNA space (5' -> 3')}

\item{probe_size}{how large should the probe be (in bp)}
}
\value{
data.frame describing probes
}
\description{
Split fusion strings into a mutant 'probe' using based on '|' breakpoint.
}
\details{
If probe size is odd - centered bases will be the first base of the downstream fusion partner
}
