% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_read_vep_txt}
\alias{probes_read_vep_txt}
\title{Title}
\usage{
probes_read_vep_txt(
  txt,
  transcript_type = c("ensembl", "refseq"),
  exclude_noncoding_rna = TRUE,
  minimal_columns = TRUE
)
}
\arguments{
\item{minimal_columns}{}
}
\description{
Title
}
