% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_write_fusion_output}
\alias{probes_write_fusion_output}
\title{Write fusion output}
\usage{
probes_write_fusion_output(fusion_sequences, outdir, prefix = "fusion_probes")
}
\arguments{
\item{fusion_sequences}{result of running \code{\link[=probes_construct_fusion_sequence]{probes_construct_fusion_sequence()}}}

\item{outdir}{Output directory}

\item{prefix}{Prefix for output files}
}
\value{
invisibly returns input fusion sequences
}
\description{
Write fusion output
}
