% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_write_output}
\alias{probes_write_output}
\title{Write output}
\usage{
probes_write_output(df_mut, outdir, prefix = "probes")
}
\arguments{
\item{df_mut}{output of probes_collapse_duplicates}

\item{outdir}{output directory}

\item{prefix}{prefix for output files}
}
\value{
Invisibly returns the same dataframe parsed
}
\description{
Write FASTQ and tabular info file
}
