% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_construct.R
\name{probes_read_vep_txt}
\alias{probes_read_vep_txt}
\title{Parse Vep annotated text file}
\usage{
probes_read_vep_txt(
  txt,
  transcript_type = c("ensembl", "refseq"),
  exclude_noncoding_rna = TRUE,
  minimal_columns = TRUE
)
}
\arguments{
\item{txt}{path to vep annotated text files}

\item{transcript_type}{Whether to use ensembl or refseq transcripts}

\item{exclude_noncoding_rna}{Should non-coding RNA mutations be excluded}

\item{minimal_columns}{Return only the minimal set of columns required}
}
\value{
data.frame describing each vep mutation
}
\description{
Read Mutations to build probes from in vep text File
}
